LPS-177665 LPS-180980 portal-search-admin-web: Add execution mode radio buttons
